+++ src/lib.rs
@@ -4,9 +4,20 @@
  `stm32f3xx-hal` contains a multi device hardware abstraction on top of the
  peripheral access API for the STMicro [STM32F3][stm] series microcontrollers.
 
+ ## Philosophie
+
+ HAL (meaning **H**ardware **A**bstraction **L**ayer) is a generic term used in many contexts,
+ but in the specific context of this crate, it is meant to abstract away the control exposed
+ by the devices "[peripheral access crate](`crate::pac`)" to simplify initialization routines,
+ with a robust interface avoiding miss-configurations while still not abstracting to much away.
+
+ Also this crate's goal is to integrate well with the rest of the rust embedded ecosystem,
+ for example by implementing the [`embedded_hal`] traits or using crates, like [`embedded_time`],
+ or [`rtcc`].
+
  [stm]: https://www.st.com/en/microcontrollers-microprocessors/stm32f3-series.html
 
- ## Basic Usagee
+ ## Basic Usage
 
  ```rust
  #![no_std]